category_service/app/requirements.txt
aniso8601==9.0.1
apispec==4.3.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autoflake==1.3
autopep8==1.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
contextlib2==0.5.5
cryptography==3.4.6
docker==4.4.4
english==2020.7.0
Flask==1.0.3
flask-apispec==0.11.0
Flask-Cors==3.0.8
Flask-Injector==0.12.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.0
idna==2.10
importlib-metadata==0.18
injector==0.18.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
loguru==0.5.3
MarkupSafe==1.1.1
marshmallow==3.0.0rc7
marshmallow-dataclass==8.3.1
marshmallow-enum==1.5.1
mccabe==0.6.1
monotonic==1.5
more-itertools==7.0.0
mypy-extensions==0.4.3
neotime==1.7.4
packaging==19.0
pansi==2020.7.3
pluggy==0.12.0
prompt-toolkit==2.0.10
py==1.8.0
py2neo==2021.0.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.8.1
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2021.1
PyYAML==5.1.1
requests==2.25.1
schema==0.7.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
wcwidth==0.1.7
webargs==7.0.1
websocket-client==0.58.0
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.1
